cleaned up school editing, removed SchoolAdmin component
